Colocate private organization UI states
