--- webapp/src/components/auth/organization/organizations-empty.tsx
@@ -1,49 +1,0 @@
-// Added with: deno task ui add @better-auth-ui/organization
-// Local changes: Replace Lucide with Phosphor icons.
-
-"use client"
-
-import { useAuthPlugin } from "@better-auth-ui/react"
-import { BriefcaseIcon as Briefcase } from "@phosphor-icons/react"
-
-import { Button } from "@/components/ui/button"
-import {
-  Empty,
-  EmptyContent,
-  EmptyDescription,
-  EmptyHeader,
-  EmptyMedia,
-  EmptyTitle,
-} from "@/components/ui/empty"
-import { organizationPlugin } from "@/lib/auth/organization-plugin"
-
-export type OrganizationsEmptyProps = {
-  onCreatePress: () => void
-  canCreate?: boolean
-}
-
-export function OrganizationsEmpty({
-  onCreatePress,
-  canCreate = true,
-}: OrganizationsEmptyProps) {
-  const { localization: organizationLocalization } = useAuthPlugin(organizationPlugin)
-
-  return (
-    <Empty>
-      <EmptyHeader>
-        <EmptyMedia variant="icon">
-          <Briefcase />
-        </EmptyMedia>
-        <EmptyTitle>{organizationLocalization.noOrganizations}</EmptyTitle>
-        <EmptyDescription>
-          {organizationLocalization.organizationsDescription}
-        </EmptyDescription>
-      </EmptyHeader>
-      <EmptyContent>
-        <Button size="sm" disabled={!canCreate} onClick={onCreatePress}>
-          {organizationLocalization.createOrganization}
-        </Button>
-      </EmptyContent>
-    </Empty>
-  )
-}
