+++ webapp/src/components/auth/organization/organizations.tsx
@@ -1,20 +1,33 @@
 // Added with: deno task ui add @better-auth-ui/organization
-// Local changes: none.
+// Local changes: Replace Lucide with Phosphor icons, activate organizations before managing them, and colocate private row and empty states.
 "use client"
 
 import type { OrganizationAuthClient } from "@better-auth-ui/core/plugins/organization"
 import { useAuth, useAuthPlugin } from "@better-auth-ui/react"
-import { useListOrganizations } from "@better-auth-ui/react/plugins/organization"
+import {
+  useListOrganizations,
+  useSetActiveOrganization,
+} from "@better-auth-ui/react/plugins/organization"
+import { BriefcaseIcon as Briefcase, GearIcon as SettingsIcon } from "@phosphor-icons/react"
+import type { Organization } from "better-auth/client"
 import { Fragment, useState } from "react"
 
 import { Button } from "@/components/ui/button"
 import { Card, CardContent } from "@/components/ui/card"
-import { Item, ItemGroup, ItemSeparator } from "@/components/ui/item"
+import {
+  Empty,
+  EmptyContent,
+  EmptyDescription,
+  EmptyHeader,
+  EmptyMedia,
+  EmptyTitle,
+} from "@/components/ui/empty"
+import { Item, ItemActions, ItemGroup, ItemSeparator } from "@/components/ui/item"
+import { Spinner } from "@/components/ui/spinner"
 import { organizationPlugin } from "@/lib/auth/organization-plugin"
 import { CreateOrganizationDialog } from "./create-organization-dialog"
-import { OrganizationRow } from "./organization-row"
 import { OrganizationViewSkeleton } from "./organization-view-skeleton"
-import { OrganizationsEmpty } from "./organizations-empty"
+import { OrganizationView } from "./organization-view"
 
 export type OrganizationsProps = {
   className?: string
@@ -100,5 +113,76 @@ export function Organizations({ className }: OrganizationsProps) {
         />
       )}
     </>
+  )
+}
+
+function OrganizationsEmpty({
+  onCreatePress,
+  canCreate = true,
+}: {
+  onCreatePress: () => void
+  canCreate?: boolean
+}) {
+  const { localization: organizationLocalization } = useAuthPlugin(organizationPlugin)
+
+  return (
+    <Empty>
+      <EmptyHeader>
+        <EmptyMedia variant="icon">
+          <Briefcase />
+        </EmptyMedia>
+        <EmptyTitle>{organizationLocalization.noOrganizations}</EmptyTitle>
+        <EmptyDescription>
+          {organizationLocalization.organizationsDescription}
+        </EmptyDescription>
+      </EmptyHeader>
+      <EmptyContent>
+        <Button size="sm" disabled={!canCreate} onClick={onCreatePress}>
+          {organizationLocalization.createOrganization}
+        </Button>
+      </EmptyContent>
+    </Empty>
+  )
+}
+
+function OrganizationRow({ organization }: { organization: Organization }) {
+  const { authClient, basePaths, navigate } = useAuth<OrganizationAuthClient>()
+  const {
+    localization: organizationLocalization,
+    viewPaths: organizationViewPaths,
+  } = useAuthPlugin(organizationPlugin)
+
+  const { mutate: setActiveOrganization, isPending: setActivePending } = useSetActiveOrganization(
+    authClient,
+    {
+      onSuccess: () => {
+        navigate({
+          to: `${basePaths.organization}/${organizationViewPaths.organization.people}`,
+        })
+      },
+    },
+  )
+
+  function manageOrganization() {
+    setActiveOrganization({ organizationId: organization.id })
+  }
+
+  return (
+    <Item>
+      <OrganizationView className="flex-1" organization={organization} />
+      <ItemActions>
+        <Button
+          variant="outline"
+          size="sm"
+          disabled={setActivePending}
+          onClick={manageOrganization}
+          aria-label={organizationLocalization.manage}
+        >
+          {setActivePending ? <Spinner /> : <SettingsIcon />}
+
+          {organizationLocalization.manage}
+        </Button>
+      </ItemActions>
+    </Item>
   )
 }
