--- webapp/src/components/auth/organization/user-invitation-row-skeleton.tsx
@@ -1,23 +1,0 @@
-// Added with: deno task ui add @better-auth-ui/organization
-// Local changes: none.
-"use client"
-
-import { Item, ItemContent, ItemMedia } from "@/components/ui/item"
-import { Skeleton } from "@/components/ui/skeleton"
-
-/**
- * Placeholder row matching `UserInvitationRow` while invitations load.
- */
-export function UserInvitationRowSkeleton() {
-  return (
-    <Item>
-      <ItemMedia>
-        <Skeleton className="size-10 shrink-0 rounded-md" />
-      </ItemMedia>
-      <ItemContent>
-        <Skeleton className="h-4 w-40 rounded-md" />
-        <Skeleton className="h-3 w-28 rounded-md" />
-      </ItemContent>
-    </Item>
-  )
-}
